Added button class, proper switch amoung states

+++ src/Common/Vec2.hpp
@@ -6,7 +6,7 @@
 
 #include <Common/Angle.hpp>
 
-template<typename T = size_t> 
+template<typename T = float> 
 class Vec2 {
 public:
     T x;
